+++ lib/potassium/assets/testing/pundit/admin_user_policy_spec.rb
@@ -1,0 +1,7 @@
+RSpec.describe BackOffice::AdminUserPolicy do
+  include_context 'with pundit setup'
+
+  let(:record) { create(:admin_user) }
+
+  it_behaves_like 'pundit back office - default permissions'
+end